Local commit

+++ src/assets/Asset.java
@@ -2,7 +2,7 @@ package assets;
 
 public interface Asset {
 
-    public float computeProfit();
+    public double computeProfit();
     public double calculateRisk(int numberOfThatItem);
